Chopping up long sourcecontext.gog
